move ToJsonPretty to App\Values, it's only meant for DTOs
