+++ app/Values/DTO.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Values;
 
-use App\Utils\ToJsonPretty;
 use Bag\Bag;
 
 abstract readonly class DTO extends Bag
